+++ src/Ydb.Sdk/src/Services/Table/ExecuteDataQuery.cs
@@ -8,6 +8,7 @@ namespace Ydb.Sdk.Services.Table;
 public class ExecuteDataQuerySettings : OperationRequestSettings
 {
     public bool KeepInQueryCache { get; set; } = true;
+    public bool AllowTruncated { get; set; }
 }
 
 public class ExecuteDataQueryResponse : ResponseWithResultBase<ExecuteDataQueryResponse.ResultData>
@@ -103,6 +104,10 @@ public partial class Session
             if (status.IsSuccess && resultProto != null)
             {
                 result = ExecuteDataQueryResponse.ResultData.FromProto(resultProto);
+                if (!settings.AllowTruncated && result.ResultSets.Any(set => set.Truncated))
+                {
+                    throw new TruncateException();
+                }
             }
 
             return new ExecuteDataQueryResponse(status, txState, tx, result);
@@ -120,4 +125,8 @@ public partial class Session
     {
         return await ExecuteDataQuery(query, txControl, new Dictionary<string, YdbValue>(), settings);
     }
+}
+
+public class TruncateException : Exception
+{
 }